Modify preprocessing to include subtypes.
A larger change here is to just symlink the types json file from the unprocessed to the preprocessed directory.  I think the old code was written the way that it was because it may have assumed function splitting, but the rest of DIRTY's code uses file splitting for which this should be fine.